Create useModifiers store hook for toolbar
